refs #1 Syncronizing with release commit: d9b773c5b4f467acbcb827a4e05d049af84e2507

+++ 3.0/src/modules/Hardware/IOPorts/src/COM/windows/SystemDeviceUtils.cpp
@@ -270,7 +270,7 @@ bool SystemDeviceUtils::enumerateSystemDevices(const QUuid & aUuid, TWinDevicePr
 
 				if (parseOK)
 				{
-					aDeviceProperties[pathProperty].VID = value;
+					aDeviceProperties[pathProperty].VID = quint16(value);
 				}
 			}
 
@@ -283,7 +283,7 @@ bool SystemDeviceUtils::enumerateSystemDevices(const QUuid & aUuid, TWinDevicePr
 
 				if (parseOK)
 				{
-					aDeviceProperties[pathProperty].PID = value;
+					aDeviceProperties[pathProperty].PID = quint16(value);
 				}
 			}
